refactor(api-credentials): isolate telemetry facts
