+++ src/services/apiCredentialProfiles/telemetry.ts
@@ -5,6 +5,7 @@ import {
 import { fetchApiCredentialModelIds } from "~/services/apiCredentialProfiles/modelCatalog"
 import { resolveApiCredentialTelemetryRequestTarget } from "~/services/apiCredentialProfiles/telemetryConfig"
 import type { TelemetryPatch } from "~/services/apiCredentialProfiles/telemetryContracts"
+import { normalizeTelemetryPatchToFacts } from "~/services/apiCredentialProfiles/telemetryFacts"
 import {
   dataLike,
   isRecord,
@@ -30,11 +31,9 @@ import { toSanitizedErrorSummary } from "~/services/verification/aiApiVerificati
 import { AuthTypeEnum, SiteHealthStatus } from "~/types"
 import type {
   ApiCredentialProfile,
-  ApiCredentialTelemetryAmount,
   ApiCredentialTelemetryAttempt,
   ApiCredentialTelemetryCapabilityMode,
   ApiCredentialTelemetryConfig,
-  ApiCredentialTelemetryFacts,
   ApiCredentialTelemetrySnapshot,
   ApiCredentialTelemetrySource,
 } from "~/types/apiCredentialProfiles"
@@ -492,149 +491,6 @@ async function querySub2ApiUsage(
         : {}),
     },
   }
-}
-
-/** Converts provider parser output into the unit-aware v6 product facts. */
-function normalizeTelemetryPatchToFacts(
-  data: TelemetryPatch,
-  source: ApiCredentialTelemetrySource,
-): ApiCredentialTelemetryFacts {
-  const facts: ApiCredentialTelemetryFacts = {}
-  const budgetSource =
-    source === API_CREDENTIAL_TELEMETRY_SOURCES.NewApiTokenUsage ||
-    source === API_CREDENTIAL_TELEMETRY_SOURCES.Sub2ApiUsage
-  const balances: NonNullable<ApiCredentialTelemetryFacts["balances"]> = []
-  const balancesToNormalize =
-    data.balances ?? (data.balance ? [data.balance] : [])
-  for (const balance of balancesToNormalize) {
-    const decimalPlaces =
-      balance.currency === TELEMETRY_PROVIDER_PROTOCOL.currencies.Jpy ? 0 : 2
-    balances.push({
-      amount: balance.amount,
-      unit: {
-        kind: "money",
-        currency: balance.currency,
-        decimalPlaces,
-      },
-      semantics:
-        source === API_CREDENTIAL_TELEMETRY_SOURCES.DeepSeekBalance
-          ? "cash"
-          : "provider-wallet",
-      ...(balance.grantedAmount !== undefined
-        ? { grantedAmount: balance.grantedAmount }
-        : {}),
-      ...(balance.toppedUpAmount !== undefined
-        ? { toppedUpAmount: balance.toppedUpAmount }
-        : {}),
-      ...(balance.isAvailable !== undefined
-        ? { isAvailable: balance.isAvailable }
-        : {}),
-    })
-  }
-
-  if (data.balanceUsd !== undefined) {
-    balances.push({
-      amount: data.balanceUsd,
-      unit: budgetSource
-        ? {
-            kind: "quota",
-            code: "usd-equivalent",
-            label: "USD-equivalent budget",
-          }
-        : {
-            kind: "money",
-            currency: TELEMETRY_PROVIDER_PROTOCOL.currencies.Usd,
-            decimalPlaces: 2,
-          },
-      semantics: budgetSource
-        ? "budget-equivalent"
-        : source === API_CREDENTIAL_TELEMETRY_SOURCES.OpenAiBilling
-          ? "cash"
-          : "legacy",
-    })
-  }
-  if (balances.length > 0) facts.balances = balances
-
-  if (data.quota) {
-    facts.quota = {
-      windows: data.quota.windows.map((window) => ({
-        type: window.type,
-        unit:
-          window.unit === "percent"
-            ? { kind: "percent" }
-            : {
-                kind: "quota",
-                code:
-                  source === API_CREDENTIAL_TELEMETRY_SOURCES.GlmQuota
-                    ? "glm-credit"
-                    : "provider-quota",
-                label:
-                  source === API_CREDENTIAL_TELEMETRY_SOURCES.GlmQuota
-                    ? "GLM credits"
-                    : "Provider quota",
-              },
-        ...(window.unit === "percent"
-          ? {}
-          : {
-              used: window.used,
-              limit: window.limit,
-              remaining: window.remaining,
-            }),
-        remainingPercent: window.percentRemaining,
-        ...(window.resetTime !== undefined
-          ? { resetTime: window.resetTime }
-          : {}),
-      })),
-      ...(data.quota.membershipLevel
-        ? { membershipLevel: data.quota.membershipLevel }
-        : {}),
-    }
-  }
-
-  const budgetUnit: ApiCredentialTelemetryAmount["unit"] = budgetSource
-    ? { kind: "quota", code: "usd-equivalent", label: "USD-equivalent budget" }
-    : {
-        kind: "money",
-        currency: TELEMETRY_PROVIDER_PROTOCOL.currencies.Usd,
-        decimalPlaces: 2,
-      }
-  const usage: NonNullable<ApiCredentialTelemetryFacts["usage"]> = {}
-  if (data.todayCostUsd !== undefined) {
-    usage.todayCost = {
-      value: data.todayCostUsd,
-      unit: {
-        kind: "money",
-        currency: TELEMETRY_PROVIDER_PROTOCOL.currencies.Usd,
-        decimalPlaces: 2,
-      },
-    }
-  }
-  if (data.todayRequests !== undefined) {
-    usage.todayRequests = {
-      value: data.todayRequests,
-      unit: { kind: "count", code: "requests" },
-    }
-  }
-  if (data.todayTokens) {
-    usage.todayTokens = {
-      ...data.todayTokens,
-      unit: { kind: "count", code: "tokens" },
-    }
-  }
-  if (data.totalUsedUsd !== undefined) {
-    usage.totalUsed = { value: data.totalUsedUsd, unit: budgetUnit }
-  }
-  if (data.totalGrantedUsd !== undefined) {
-    usage.totalGranted = { value: data.totalGrantedUsd, unit: budgetUnit }
-  }
-  if (data.totalAvailableUsd !== undefined) {
-    usage.totalAvailable = { value: data.totalAvailableUsd, unit: budgetUnit }
-  }
-  if (data.expiresAt !== undefined) usage.expiresAt = data.expiresAt
-  if (data.unlimitedQuota !== undefined) usage.unlimited = data.unlimitedQuota
-  if (Object.keys(usage).length > 0) facts.usage = usage
-
-  return facts
 }
 
 /**
